requirements.txt
# Python requirements for OCR PDF extraction
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0
